% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_accuracy.R
\name{check_data_accuracy}
\alias{check_data_accuracy}
\title{Check data accuracy}
\usage{
check_data_accuracy(df)
}
\arguments{
\item{df}{A data frame to be checked.}
}
\value{
A message indicating if extreme values are found or not.
}
\description{
This function checks the accuracy of the data by looking for outliers or extreme values.
}
\examples{
df <- data.frame(x = c(1,2,3,4,5), y = c(2,4,6,8,10))
check_data_accuracy(df)

}
