% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_quality.R
\name{check_data_quality}
\alias{check_data_quality}
\title{Check the quality of a data frame}
\usage{
check_data_quality(df)
}
\arguments{
\item{df}{A data frame to be checked}
}
\value{
A list of results from each check
}
\description{
This function performs a series of checks on a data frame to assess its overall quality, including its structure, missing values, accuracy, distribution, reliability, negative values, outliers, sample size, and duplicate rows/columns.
}
\examples{
df <- data.frame(x = 1:10, y = c(1, 2, NA, 4, 5, 6, 7, 8, 9, 10))
check_data_quality(df)

}
