% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_structure.R
\name{check_data_structure}
\alias{check_data_structure}
\title{Check the structure of the data}
\usage{
check_data_structure(df)
}
\arguments{
\item{df}{The data frame to be checked.}
}
\value{
None
}
\description{
This function checks the structure of the given data frame and prints the number of rows,
number of columns, column names, column data types, and number of missing values.
}
\examples{
check_data_structure(my_data)

}
