% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_negative_values.R
\name{check_for_negative_values}
\alias{check_for_negative_values}
\title{Check if a data frame contains negative values.}
\usage{
check_for_negative_values(df)
}
\arguments{
\item{df}{The data frame to check for negative values.}
}
\value{
If negative values are found, the function returns their indices as an array index object.
If no negative values are found, NULL is returned.
}
\description{
This function checks if a data frame contains negative values and returns their indices if any are found.
}
\examples{
df <- data.frame(a = c(1, 2, 3), b = c(-1, 0, 1))
check_for_negative_values(df)
# [1] "Data frame contains negative values."
#      row col
# [1,]   2   1"

}
