% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_values.R
\name{check_missing_values}
\alias{check_missing_values}
\title{Check for Missing Values in Data Frame}
\usage{
check_missing_values(df)
}
\arguments{
\item{df}{A data frame to check for missing values.}
}
\value{
A message indicating if missing values were found or not.
}
\description{
This function checks for missing values in a data frame and prints out the names of the columns with missing values and their counts.
}
\examples{
df <- data.frame(x = c(1,2,3), y = c(NA,5,6))
check_missing_values(df)

}
\author{
Your Name
}
