% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_outliers}
\alias{check_outliers}
\title{Check for Outliers or Extreme Values in Data}
\usage{
check_outliers(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A message indicating whether or not extreme values were found.
}
\description{
This function checks for outliers or extreme values in a given dataframe.
}
\examples{
check_outliers(mtcars)

}
