% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sample_size.R
\name{check_sample_size}
\alias{check_sample_size}
\title{Check if sample size is adequate}
\usage{
check_sample_size(df)
}
\arguments{
\item{df}{A data frame to be checked}
}
\value{
A message indicating if the sample size is adequate or too small
}
\description{
This function checks if the sample size of a data frame is adequate for statistical analysis.
}
\examples{
df <- data.frame(x = rnorm(50), y = rnorm(50))
check_sample_size(df)

}
