% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.packages.R
\name{find.packages}
\alias{find.packages}
\title{Find R packages that can import a file format}
\usage{
find.packages(file_extension)
}
\arguments{
\item{file_extension}{The file extension for the file format to search for packages to import}
}
\value{
A character vector of package names that can be used to import the file format
}
\description{
This function searches the CRAN repository for R packages that can be used to import a file format
}
