% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.packages_path.R
\name{find.packages_path}
\alias{find.packages_path}
\title{Find the R Packages to Import a File Format}
\usage{
find.packages_path(file_path)
}
\arguments{
\item{file_path}{A character string specifying the file path of the file to be imported.}
}
\value{
A character string that lists the R packages that can be used to import the file format of the input file.
}
\description{
This function takes a file path as input and searches the CRAN repository for R packages that can import the file format.
}
\examples{
# Search for packages that can import a CSV file
find.packages_path("sample.csv")

# Search for packages that can import a JSON file
find.packages_path("sample.json")

}
