% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.data.R
\name{fix.data}
\alias{fix.data}
\title{Fix data frame column and row names and remove symbols and blanks}
\usage{
fix.data(df)
}
\arguments{
\item{df}{A data frame to be processed.}
}
\value{
The cleaned data frame.
}
\description{
This function applies several data cleaning functions from the \code{fixr} package to a given data frame. The \code{fix_data_names}, \code{remove_spaces}, \code{remove_symbols_data}, and \code{replace_blanks_with_na} functions are used to add "X_" before column and row names that start with a number, remove leading/trailing spaces, remove non-alphanumeric characters from the data, replace spaces with underscores in column and row names, and replace empty string values with NAs, respectively.
}
\examples{
df <- data.frame(" 1st col " = c("", "foo", ""), "2nd col" = c(" ", " ", "bar"),
                 "3rd col" = c(1, 2, 3))
fix.data(df)

}
