% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_blanks_with_na.R
\name{fix_blanks_with_na}
\alias{fix_blanks_with_na}
\title{Replace blanks with NA in a data frame}
\usage{
fix_blanks_with_na(df)
}
\arguments{
\item{df}{A data frame to be processed.}
}
\value{
The data frame with empty string values replaced with NAs.
}
\description{
This function replaces all empty string values ("") in a given data frame with NA values.
}
\examples{
df <- data.frame(x = c("", "foo", ""), y = c("", "", "bar"), z = c(1, 2, 3))
fix_blanks_with_na(df)

}
