% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_column_names.R
\name{fix_col_names}
\alias{fix_col_names}
\title{Fix Data Frame Column Names}
\usage{
fix_col_names(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A modified data frame with updated column names
}
\description{
This function adds "X_" before the names of any columns in a data frame
that start with a number, and removes any leading or trailing spaces in the column
names. This is useful to avoid errors that can occur when
working with column names in some R functions.
}
\examples{
df <- data.frame(" 1st column " = 1:5, "2nd column" = 6:10, " third column " = 11:15)
fix_col_names(df)

# Output:
# X_1st_column X_2nd_column third_column
# 1 1 6 11
# 2 2 7 12
# 3 3 8 13
# 4 4 9 14
# 5 5 10 15

}
