% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_data_names.R
\name{fix_data_names}
\alias{fix_data_names}
\title{Fix Data Frame Column and Row Names}
\usage{
fix_data_names(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A modified data frame with updated column and row names
}
\description{
This function adds "X_" before the names of any columns or rows in a data frame
that start with a number, and removes any leading or trailing spaces in the column
and row names. This is useful to avoid errors that can occur when
working with column or row names in some R functions.
}
\examples{
df <- data.frame(" 1st column " = 1:5, "2nd column" = 6:10, " third column " = 11:15)
fix_data_names(df)

# Output:
#   X_1st_column X_2nd_column third_column
# 1            1             6           11
# 2            2             7           12
# 3            3             8           13
# 4            4             9           14
# 5            5            10           15

}
