% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_and_F.R
\name{fix_dir_T_and_F}
\alias{fix_dir_T_and_F}
\title{Fix T and F Directory}
\usage{
fix_dir_T_and_F(path = "R")
}
\arguments{
\item{path}{folder to clean}
}
\value{
invisible NULL
}
\description{
Fix T and F Directory
}
\examples{
# TODO
}
