% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_outliers.R
\name{fix_outliers}
\alias{fix_outliers}
\title{Remove Outliers from a Data Frame}
\usage{
fix_outliers(df)
}
\arguments{
\item{df}{A data frame to clean}
}
\value{
A cleaned data frame with outlier rows removed
}
\description{
This function removes outlier rows from a data frame by identifying rows with
values that are more than 2 standard deviations away from the mean in any column.
}
\examples{
df <- data.frame(x = c(1,2,3,4,5,6,7,8,9,10),
                 y = c(1,1,1,1,1,1,1,100,1,1))
fix_outliers(df)

}
