% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_row_names.R
\name{fix_row_names}
\alias{fix_row_names}
\title{Fix Data Frame Row Names}
\usage{
fix_row_names(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A modified data frame with updated row names
}
\description{
This function adds "X_" before the names of any rows in a data frame
that start with a number, and removes any leading or trailing spaces in the row
names. This is useful to avoid errors that can occur when
working with row names in some R functions.
}
\examples{
df <- data.frame(" 1st column " = 1:5, "2nd column" = 6:10, " third column " = 11:15)
row.names(df) <- c(" 1st row ", "2nd row", " third row ")
fix_row_names(df)

# Output:
# X_1st_column X_2nd_column third_column
# X_1st_row 1 6 11
# 2nd_row 2 7 12
# X_third_row 3 8 13
# 4 4 9 14
# 5 5 10 15

}
