% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_and_F.R
\name{fix_T_and_F}
\alias{fix_T_and_F}
\title{Fix T and F Files}
\usage{
fix_T_and_F(path)
}
\arguments{
\item{path}{file to clean}
}
\value{
invisble NULL
}
\description{
Fix T and F Files
}
\examples{
# TODO
}
