% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_spaces.R
\name{remove_spaces}
\alias{remove_spaces}
\title{Replace spaces in column and row names with underscores}
\usage{
remove_spaces(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A modified data frame with spaces in column and row names replaced by underscores.
}
\description{
This function takes a data frame as an argument and replaces all spaces in
the column names and row names with underscores.
}
\examples{
my_data <- data.frame("Column Name 1" = c(1, 2, 3), "Column Name 2" = c(4, 5, 6))
rownames(my_data) <- c("Row Name 1", "Row Name 2", "Row Name 3")

remove_spaces(my_data)
# Returns a data frame with column and row names where spaces are replaced by underscores.

}
