% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_symbols_data.R
\name{remove_symbols_data}
\alias{remove_symbols_data}
\title{Remove non-alphanumeric characters from data frame}
\usage{
remove_symbols_data(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A modified data frame with all non-alphanumeric characters removed
from the data, but not from the column or row names.
}
\description{
This function takes a data frame as an argument and removes all characters
other than numbers and alphabets from the data, but not from the column names
or row names. The function also replaces spaces in column and row names with
underscores.
}
\examples{
my_data <- data.frame("Column Name 1" = c("1", "2$", "3*"), "Column Name 2" = c("4", "5(", "6)"))
rownames(my_data) <- c("Row Name 1", "Row Name 2", "Row Name 3")

remove_symbols_data(my_data)
# Returns a modified data frame with non-alphanumeric characters removed
from the data, but not from the column or row names.

}
