% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_quality.R
\name{check_data_quality}
\alias{check_data_quality}
\title{Check Data Quality}
\usage{
check_data_quality(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A message indicating the results of each data quality check.
}
\description{
This function performs a series of data quality checks on a given dataframe, including checking the data structure, missing values, data accuracy, negative values, outliers, sample size, duplicate rows, and duplicate columns.
}
\examples{
df <- data.frame(w = c(7, 8, 180, 7), x = c("a", "b", "c", "a"),
                 y = c(4, NA, -6, 4), z = c(7, 8, 180, 7))

# Check the data quality of the example dataframe
check_data_quality(df)

}
