% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_structure.R
\name{check_data_structure}
\alias{check_data_structure}
\title{Check the structure of the data}
\usage{
check_data_structure(df)
}
\arguments{
\item{df}{The data frame to be checked.}
}
\value{
None
}
\description{
This function checks the structure of the given data frame and prints the number of rows,
number of columns, column names, column data types, and number of missing values.
}
\examples{
df <- data.frame(id = 1:10,
gender = c("male", "female", "male", "male", "male", "male", "male", "male", "female", "female"),
age = c(25, 32, 45, 19, 27, 56, 38, 42, 33, NA),
salary = c(50000, 60000, 75000, 45000, 55000, 90000, NA, 80000, 65000, 70000))

# Check the data structure of the example dataframe
check_data_structure(df)

}
