% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sample_size.R
\name{check_sample_size}
\alias{check_sample_size}
\title{Check if sample size is adequate}
\usage{
check_sample_size(df)
}
\arguments{
\item{df}{A data frame to be checked}
}
\value{
A message indicating if the sample size is adequate or too small
}
\description{
This function checks if the sample size of a data frame is adequate for statistical analysis.
}
\examples{
df <- data.frame(w = c(7, 8, 180, 7), x = c("a", "b", "c", "a"),
                 y = c(4, 5, -6, 4), z = c(7, 8, 18, 7))
check_sample_size(df)

}
